Added ability to use the Barlow Twins Loss
Barlow Twins Loss encoded as CrossCorrLoss, requires both the
representation size and the lambda value (determines how much weighting
should be given to making the off-diagonal terms 0).

Several new command line arguments were created to allow to use the
CrossCorrLoss. `latent_loss` allows you to select either 'mse' for
MSELoss or 'covar' for the CrossCorrLoss. `crosscorr_lambda` to
determine the weighting of the off-diagonal loss, default value is the
value used in the Barlow Twins paper.